fix(http2): Fixed GOAWAY race for 'error' event.
It seems that after a session GOAWAY, the stream can get an 'error'
event instead of 'frameError' to be notified of this.
There is therefore no proper error code signalling exactly what
happened, so an error message string is being matched. Not ideal...

Potentially, if the session keeps a set of its streams, the GOAWAY
could be forwarded to the streams, but that's a lot more complex.